requirements.txt
pypdf
transformers
sentence-transformers
faiss-cpu
torch # sentence-transformers depends on torch
numpy
